feat: replaced dummy analytics page 🚀
